Fixed GitHub Repository
I haven't been able to commit anything to the master branch in my repository the entire time that I've been working on this project and I never understood why. I had issues in the beginning of the semester with my log-in and having to use other people's log-ins on the computer. Storm helped me to merge the repositories that I had (they had the same files). So that I could finally commit to the master branch of the proper repository.